chtMultiRegionFoam: Added optional thermal system sub-iteration
For some systems the thermal coupling between the solid and fluid regions
dominates overall convergence and it may be beneficial to sub-iterate over the
thermal system as the energy solution is cheaper than the pressure-velocity
system.  To test this the new optional nEcorr entry in system/fvSolution is
provided which defaults to 1, e.g.

PIMPLE
{
    nOuterCorrectors 1;
    nEcorr           2;
}

If this proves useful it could be extended and improved by adding convergence
controls.

Note that the solution of the solid regions is now before the fluid regions as
it make more sense if solid regions are sources or sinks of heat to solve them
before the fluid.