setAtmBoundaryLayer: New initialisation utility
This utility initialises the U, k and epsilon fields (if available) to
the output of the atmBoundaryLayer model. This is the same model as used
in the atmBoundaryLayerInlet.* boundary conditions and in the
waveAtmBoundaryLayer wave model.

The settings for the initialisation are read from a
system/setAtmBoundaryLayerDict file and are identical to the settings
required by the other use cases. An example of the settings required
within a system/setAtmBoundaryLayerDict file is shown below:

    zDir            (0 0 1);        // Vertical direction
    flowDir         (1 0 0);        // Direction of far-field flow
    Zref            20;             // Reference height
    Uref            10;             // Speed at reference height
    z0              uniform 0.1;    // Roughness height
    zGround         uniform 0;      // Ground height