multiphaseEulerFoam: Multiphase blending
Blending of sub-models has been extended so that it provides the
necessary functionality for the simulation of three or more phases.

Models now, by default, blend within the two-phase subset. So, the
functions which specify blending coefficients are evaluated using
`alpha1/(alpha1 + alpha2)` and `alpha2/(alpha1 + alpha2)`, rather than
just `alpha1` and `alpha2`. This ensures that the functions behave
consistently in multiphase configurations as the combined fraction of
both phases reduces.

Additional "displaced" categories of model are now available. These
allow the interface modelling to change as a third phase fraction
becomes significant and displaces the phases associated with the
interface in question. These can be specified using displaced phase
interfaces. If they are not supplied then the non-displaced models apply
instead.

Additional "general" categories are also now available. These span the
entire range of the subset's phase fraction space and apply if models
for the more specific configurations (i.e., displaced and segregated)
are omitted.

For example, to specify standard SchillerNaumann droplet/bubble drag
modelling outside of a solid bed, and the bed-specific AttouFerschneider
drag model within the bed, the following specification could be used:

    drag
    {
        air_dispersedIn_water
        {
            type            SchillerNaumann;
            residualRe      1e-3;
        }

        water_dispersedIn_air
        {
            $air_dispersedIn_water;
        }

        air_water_displacedBy_solid
        {
            type            AttouFerschneider;
            gas             air;
            liquid          water;
            solid           solid;
            E1              280;
            E2              4.8;
        }

        solid_dispersedIn_air
        {
            $air_water_displacedBy_solid;
        }

        solid_dispersedIn_water
        {
            $air_water_displacedBy_solid;
        }
    }

The "air_water_displacedBy_solid" model shown above is a
"general-displaced" model, in that its application does not depend on
dispersal of the air-water system; it only depends on the level of
displacement by the solid. Alternatively, a set of three
"dispersed-displaced" and "segregated-displaced" models could have been
provided with the following syntax:

        air_dispersedIn_water_displacedBy_solid
        {
            ...
        }

        air_segregatedWith_water_displacedBy_solid
        {
            ...
        }

        water_dispersedIn_air_displacedBy_solid
        {
            ...
        }

The blending methods themselves have changed slightly. There are now two
new methods; "segregated" for churning configurations in which no phase
is ever considered dispersed in another, and "continuous" where a
specified phase is always considered to be continuous and all other
phases are considered dispersed. These models replace the "none" model.

The linear and hyperbolic blending methods now support a "none" entry,
which can be set for parameters for phases that cannot become
continuous. For example, in an air-water-particles simulation the air
and water may become continuous, but the particles may not. A suitable
linear blending specification for this scenario might be as follows:

    blending
    {
        default
        {
            type            linear;
            minFullyContinuousAlpha.air 0.8;
            minPartlyContinuousAlpha.air 0.2;
            minFullyContinuousAlpha.water 0.8;
            minPartlyContinuousAlpha.water 0.2;
            minFullyContinuousAlpha.solid none;
            minPartlyContinuousAlpha.solid none;
        }
    }

The blending now provides far more feedback in terms of warnings and
errors when invalid or inadvisable specifications are made.

It is now also possible to visualise the various models' utilisation of
the phase fraction space. If the keyword "format" is placed in the
blending section of constant/phaseProperties then a plot (1 or 2 phases)
or surface (3 or more phases) file will be written out which shows the
phase fractions and the coefficients with which various models are
combined.