FvFaceCellWave: Alternative to FaceCellWave based on FV connectivity
This avoids the need to explicity handle differences in how the polyMesh
relates to the fvMesh (e.g., empty patches) when building a result that
relates to the fvMesh. This is now used for waving data associated with
the finite volume mesh; e.g., for computing wall distances or smoothing.